requirements.txt
# # openai-agents==0.2.10
# python-dotenv==1.1.1
# playwright==1.54.0
# # browser-use==0.7.1
# Crawl4AI==0.7.4
# # openai==1.99.5
# fastapi==0.116.1
# # litellm==1.76.1
# Flask==3.1.2
# flask-cors>=4.0.0

openai-agents==0.2.11
playwright==1.55.0
python-dotenv==1.1.1
aiohttp==3.12.15
browser-use==0.7.7
Crawl4AI==0.7.4
tabulate==0.9.0
Flask-RESTful==0.3.10
Flask==3.1.2
flask-cors==6.0.1
PyJWT==2.10.1
Quart==0.20.0
quart-cors==0.8.0